Using `transpileModule` instead of `createProgram` fixed the problem with slow builds on a bigger codebase. 
Was testing on a set of 660 ts/tsx files.
Based on a documentation from [TS wiki](https://github.com/Microsoft/TypeScript/wiki/Using-the-Compiler-API#transpiling-a-single-file).

## Performance results

> Test Suites: 660 passed, 660 total
> Tests:       660 passed, 660 total
> Snapshots:   0 total

### With this PR

> Time:        3.329s, estimated 12s
> Ran all test suites.
> ✨  Done in **4.77s**.

### Current v1.4.0

Around **5 min**

### Actual `tsc --noEmit`

Around **12.59s**